Add files via upload
This database is ideal for small zoo. There are a lot of functionalities, which will save a lot of your time. For example there are 7 queries, using which you will be able to do complicated operations in one click. This queries will allow you to see health of your animals, producer of fodder, see the salaries of staff, delete empty tables and connect two tables in one. Moreover there is some Modules for easier navigation and macros. There is also a report, where you can find almost all about your animals.